% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_by_group.R
\name{summary_by_group}
\alias{summary_by_group}
\title{Summary Statistics of Numerical Variable by Group}
\usage{
summary_by_group(numeric_vector, group_vector)
}
\arguments{
\item{numeric_vector}{An vector of class "numeric", each numerical value is a part of one observation row.
To make the name of parameter be more specific, I named it as "numeric_vector".}

\item{group_vector}{An vector which represents the group type of each numeric value in the "numeric_vector" one by one.
To emphasize the group type, I named it as "group_vector".}
}
\value{
An object of class "data.frame" that contains the summary statistics of each group type.
}
\description{
Get the summary statistics (max, min, mean and median) of a numerical variable
across the groups of a categorical variable.
}
\examples{
summary_by_group(c(1.5, 7.6, 3.0, 4.7, 8, 8.4), c("A", "C", "A", "B", "B", "A"))
summary_by_group(datateachr::vancouver_trees$diameter, datateachr::vancouver_trees$genus_name)
}
